Make ServiceNameDetector and interface
